Apply Double type on seekTo() and getCurrentTime()
- milliseconds 계산을 위해 int에서 double로 type 변경

+++ android/src/main/java/com/inprogress/reactnativeyoutube/YouTubePlayerController.java
@@ -162,7 +162,7 @@ public class YouTubePlayerController implements
     }
 
     public double getCurrentTime() {
-      return mYouTubePlayer.getCurrentTimeMillis() / 1000;
+      return (double)mYouTubePlayer.getCurrentTimeMillis() / 1000.0;
     }
 
     public int getDuration() {
